+++ SourceCode/e_commerce_app/lib/screens/screens.dart
@@ -1,0 +1,1 @@
+export 'package:e_commerce_app/screens/home/home_screen.dart';
